TS: SimpleCommonComponents/HighlightedText
