+++ src/components/SimpleCommonComponents/HighlightedText/test.js
@@ -1,4 +1,3 @@
-// @flow
 import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
