+++ src/pages/Proteome/index.js
@@ -18,6 +18,7 @@ import Table, {
 import File from 'components/File';
 // $FlowFixMe
 import Tooltip from 'components/SimpleCommonComponents/Tooltip';
+// $FlowFixMe
 import HighlightedText from 'components/SimpleCommonComponents/HighlightedText';
 import NumberComponent from 'components/NumberComponent';
 import MemberDBSelector from 'components/MemberDBSelector';
